requirements.txt
dnspython==2.7.0
pymongo==4.12.0
python-dotenv==1.1.0
pytest==8.3.5
requests==2.32.3
fastapi==0.115.12
uvicorn==0.34.1
motor==3.7.0
pydantic==2.11.3
allure-pytest==2.14.0